Refactor buildDocs.js into separate steps
The original buildDocs.js was very opinionated about what documents
are being generated.  This led to a rigid workflow that could not
be repurposed for generating other documents.  We want to model/
generate the model data structures used to capture DOF, so the
rigid workflow had to be made flexible.  The refactored scripts
to provide the desired flexibility.  The new workflow is:

 - Create/update architecture model files (per spec to be modeled)

 - Run buildUnifiedModel.js to process the architecture model files
   into a single unified model yaml file to be used in further
   steps in the workflow

 - Create (or update as needed) liquid templates for generated documents

 - Run generateDoc using the unified model for each template to
   generate each desired document

Repeat the above process when changes are made to the architecture
models or templates.  The shell script buildDOF.sh demonstrates
automation of this process.